Simplify `selection_engines` by centering around a single Selection() class.
