+++ pypunisher/__init__.py
@@ -1,2 +1,2 @@
 from pypunisher.metrics.criterion import aic, bic
-from pypunisher.selection_engines.unified import Selection
+from pypunisher.selection_engines.selection import Selection
