+++ pypunisher/selection_engines/selection.py
@@ -1,7 +1,7 @@
 """
 
-    Forward Selection Algorithm
-    ~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    Forward and Backward Selection Algorithms
+    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 """
 from pypunisher.selection_engines._utils import (get_n_features,
@@ -12,14 +12,14 @@ from pypunisher.selection_engines._utils import (get_n_features,
                                                  input_checks)
 
 
-class ForwardSelection(object):
-    """Forward Selection Algorithm.
+class Selection(object):
+    """Unified Forward and Backward Selection Class.
 
     Args:
         model : sklearn model
             any sklearn model with `.fit()`, `.predict()` and
             `.score()` methods.
-        X_train ndarray:
+        X_train : ndarray
             a 2D numpy array of (observations, features).
         y_train : ndarray:
             a 1D array of target classes for X_train.
@@ -36,6 +36,11 @@ class ForwardSelection(object):
         verbose : bool
             if True, print additional information as selection occurs.
             Defaults to True.
+
+    Exposes:
+        * .forward()
+        * .backward()
+
     """
 
     def __init__(self, model, X_train, y_train,
@@ -117,6 +122,78 @@ class ForwardSelection(object):
 
             if self._forward_break_criteria(S, j_score_dict=j_score_dict,
                                             max_features=max_features):
+                break
+
+        return S
+
+    def _fit_and_score_back(self, S, exclude):
+        features = [f for f in S if f != exclude] if exclude else S
+        self._model.fit(self._X_train[:, features], self._y_train)
+        return self._model.score(self._X_val[:, features], self._y_val)
+
+    def backward(self, n_features=0.5, min_change=None):
+        """Run Backward Selection Algorithm.
+
+        Args:
+            n_features : int or float
+                The number of features to select.
+                Floats will be regarded as proportions of the total
+                that must lie on (0, 1).
+                `min_change` must be None for `n_features` to operate.
+            min_change : int or float, optional
+                The smallest change to be considered significant.
+                `n_features` must be None for `min_change` to operate.
+
+        Returns:
+            S : list
+              The column indices of `X_train` (and `X_val`) that denote the chosen features.
+
+        Raises:
+            if `n_features` and `min_change` are both non-None.
+
+        """
+        input_checks(locals())
+        S = list(range(self._n_features))  # start with all features
+
+        if n_features:
+            n_features = parse_features_param(
+                n_features, total=len(S), param_name="n_features"
+            )
+
+        last_score = self._fit_and_score_back(S, exclude=None)
+        worse_case = worse_case_bar(self._n_features, verbose=self._verbose)
+        for _ in worse_case:
+            worse_case.set_postfix(n_features=len(S), score=last_score)
+
+            # 1. Hunt for the least predictive feature.
+            best = None
+            for j in S:
+                score = self._fit_and_score_back(S, exclude=j)
+                if score >= last_score and (best is None or score > best[1]):
+                    best = (j, score)
+
+            if not best:
+                break  # Relent. Removing any `j` yielded a lower score.
+            else:
+                to_drop, best_new_score = best
+
+            # 2a. Halting Based Blindly Based on `n_features`.
+            if isinstance(n_features, int):
+                S.remove(to_drop)  # blindly drop.
+                last_score = best_new_score
+                if len(S) == n_features:
+                    break
+                else:
+                    continue  # i.e., ignore criteria below.
+
+            # 2b. Halt if the change is not longer significant.
+            if (best_new_score - last_score) < min_change:
+                break
+            else:
+                S.remove(to_drop)
+                last_score = best_new_score
+
+            if len(S) == 1:  # continuing is futile.
                 break
 
         return S
